consolidate job queue utilities
