--- orchestrator/__init__.py
@@ -1,7 +1,0 @@
-"""Compatibility shims for moved worker modules."""
-
-from .scheduler import scheduler  # noqa: F401
-from .worker import start_worker  # noqa: F401
-from . import worker  # noqa: F401
-
-__all__ = ["scheduler", "start_worker", "worker"]
