--- orchestrator/worker.py
@@ -1,3 +1,0 @@
-from workers.worker import *  # noqa: F401,F403
-
-__all__ = [name for name in globals() if not name.startswith("_")]
