+++ pyproject.toml
@@ -13,6 +13,4 @@ known_first_party = [
   "platform",
   "analysis",
   "storage",
-  "orchestrator",
-  "workers",
 ]
