+++ utils/__init__.py
@@ -1,9 +1,10 @@
 """Shared utility subpackages."""
 
-from . import display_utils, logging_utils, reporting_utils
+from . import display_utils, logging_utils, reporting_utils, jobqueue
 
 __all__ = [
     "display_utils",
     "logging_utils",
     "reporting_utils",
+    "jobqueue",
 ]
