--- workers/__init__.py
@@ -1,5 +1,0 @@
-"""Simple orchestration utilities for analysis jobs."""
-from .scheduler import scheduler
-from .worker import start_worker
-
-__all__ = ["scheduler", "start_worker"]
